GL-9 Infinite Uptime
Currently the app.py will now run until cancelled, sleeping every 3 seconds.
I had to add a conditional if isinstance(data,Account) within database.managers.py to ensure that if the account information has already
been added to the database from a previous iteration of main() it would not run into a unique_key error since it would already exist.
It seems to run correctly. The first instance steps in the conditional isinstance statement
(represented by a print statement on line 177 in managers.py) and
bash: line 1: M-F: command not found
does not step in on subsequent iterations (represented by a lack of print statement).
This is also gated by a .get.account conditional. I am unsure if the get.account
is working as intended since it does not give me a print("step in check") on line 54 in managers.py